+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Render activity, attendance and savings for each subscription.
+- Bottom sheet to rate the app.
 
 ## [2024.12.2] - 2024-12-20